refs #9245 not logged users can add files; add transfer limits for users

+++ db/migrate/20170102152920_add_max_upload_size_to_users.rb
@@ -1,5 +1,5 @@
 class AddMaxUploadSizeToUsers < ActiveRecord::Migration
   def change
-    add_column :users, :max_upload_size, :decimal, default: 10
+    add_column :users, :max_upload_size, :float, default: 10.0
   end
 end
